refactor : 파일 위치 변경, 폴더이름 변경 및 index.jsx 삭제
- 폴더이름 /frontend => /client
- 불필요한 components/index.jsx 파일 제거
- /pages 의 About, Home 위치 이동 => /components
