+++ client/src/components/About/About.jsx
@@ -3,13 +3,13 @@ import React from 'react';
 const About = () => {
     return (
         <div className="container px-3">
-            <br/>
-                <h2>About</h2>
-                <hr/>
+            <br />
+            <h2>About</h2>
+            <hr />
                 금오공과대학교 컴퓨터소프트웨어공학과 백준 문제풀이 스코어보드입니다.
-                <br/>
+            <br />
                 문제를 풀고난 후 월요일 이전에는 반드시 [Home] - [갱신] 버튼을 눌러주시기 바랍니다 !!
-            <br/>
+            <br />
         </div>
     );
 };
