+++ client/src/components/Home/Update.jsx
@@ -1,4 +1,4 @@
-import React,{useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 import Button from 'react-bootstrap/Button';
 import Alert from 'react-bootstrap/Alert';
 import { updateSolvedProblem } from '../../store/ApiStore';
@@ -8,7 +8,7 @@ const Update = (props) => {
     const [isSuccess, setIsSuccess] = useState(false);
     const [isFail, setIsFail] = useState(false);
     const reloadData = props.reloadData;
-    
+
     useEffect(() => {
         handleLoading();
     }, []);
@@ -23,7 +23,7 @@ const Update = (props) => {
         await setIsFail(false);
         const res = await updateSolvedProblem();
         await setIsLoading(false);
-        if(res === "SUCCESS") {
+        if (res === "SUCCESS") {
             setIsSuccess(true);
             reloadData()
         }
@@ -34,7 +34,7 @@ const Update = (props) => {
 
     const handleLoading = () => {
         if (isLoading) {
-            setLoadingTime().then(() => {setIsLoading(false);});
+            setLoadingTime().then(() => { setIsLoading(false); });
         }
     }
 
@@ -48,20 +48,20 @@ const Update = (props) => {
             >
                 {isLoading ? 'Loading…' : '갱신'}
             </Button>
-        <br/><br/>
-            <Alert 
+            <br /><br />
+            <Alert
                 variant="success"
                 show={isSuccess ? true : false}
             >
                 갱신 성공!
             </Alert>
-            <Alert 
+            <Alert
                 variant="danger"
                 show={isFail ? true : false}
             >
                 갱신 실패!
             </Alert>
-        <br/><br/>
+            <br /><br />
         </>
     );
 };
