+++ client/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDom from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
-import { App } from './components/index';
+import App from './components/App';
 
 import './index.css';
 import 'bootstrap/dist/css/bootstrap.css';
