--- frontend/src/components/index.jsx
@@ -1,2 +1,0 @@
-export { default as NavBar } from './NavBar';
-export { default as App } from './App';
